requirements.txt
#requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
python-dotenv
openai
twilio
python-multipart

